making the salary a float instead of integer

+++ Assignment1.py
@@ -1,4 +1,4 @@
-salary=input(int("Please enter your salary for the month: "))
+salary=input(float("Please enter your salary for the month: "))
 month=input("Please enter the name of the month: ")
 savingsPercentage=input(float("Please enter the percentage for your savings: "))
 rentPercentage=input(float("Please enter the percentage for your rent: "))
